requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
matplotlib>=3.4.0
pillow>=8.0.0
scikit-image>=0.18.0
tqdm>=4.62.0
pyyaml>=5.4.0
tensorboard>=2.8.0
pytest>=6.2.0
jupyter>=1.0.0 